bin: use ghauth for fetching token
This is a slight modification to the bin to use `ghauth` to simplify
the process of fetching / creating a token. I opted to put it in bin
rather than lib due to simplicity. The current architecture of lib/,
specifically having the github object being created in the constructor
does not play especially nice with the callback style of `ghauth`.

Auth happening in lib would probably be nicer, but I bet very few
will use this as a client lib so ¯\_(ツ)_/¯